requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
torchvision==0.16.0
opencv-python-headless==4.8.1.78
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
albumentations==1.3.1
tqdm==4.66.1
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-logging==0.42b0
opentelemetry-instrumentation-system-metrics==0.42b0
opentelemetry-exporter-otlp-proto-http==1.21.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.25.2
